Create LLC and TSC (#2107)
* Create LLC and TSC

Clean up various legal/charter constructs.

The Linux Foundation plans to move this project into its own LLC
(this provides various legal protections), and we must state that.

A side-effect is that we need to create "Technical Steering Committee"
(TSC). I think that's a good thing. I intend to start it with me
(as I was the original project lead & committer), and work to add
others to it, so it's not just me working by myself.

The proposed approach says the OpenSSF TAC can change the TSC
by 2/3rds vote. The intent is to ensure that I (or anyone else) can't
be a dictator, and that it's possible to get things going again
if something goes wrong. Originally I was going to have the
OpenSSF Best Practices WG do that vote, but I think it'd be
better to have the TAC do it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add CRob to proposed TSC list

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Small tweaks to TSC/governance materials

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Add Tony Hansen to TSC member list

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak link text so it's natural language text

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix tests using 'Projects' as text

Capybara is over-matching on the term "Projects", leading to ambiguous
links. The documentation isn't clear on whether or not 'exact: true'
is supported in this context, but that's the obvious solution. so try
that first.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak test to properly perform exact match

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Create simple way to disambiguate Projects links

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Change some tests to use basic find() to avoid ambiguity

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Minor clarifications

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

---------

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>