Add section ids to project badge entry
Add ids for every section, to make it easy to jump directly to them.
Section ids have the form "section_" followed by the English name
of the section downcased and with spaces turned into underscores.
Only characters A-Za-z0-9_- are used, any other characters are removed.

The "section_" prefix is necessary because there are cases where the
section name and the criterion overlap (e.g., `floss_license`).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>