Delayed job to purge CDN data to fix race condition (#2130)
Add a delayed job to purge CDN project data to fix
a race condition.

In some cases the CDN can end up with old data due to
a race condition between the server and CDN. There doesn't seem
to be a reasonable solution to *end* the race. So to solve this,
add an additional delayed job to re-purge data for a given
project. That way, if the CDN loses the race, the obsolete data
will soon be purged and the next request will get the correct data.
At worst there will be a few seconds where users see the old data,
but that's unsurprising behavior - it will look like it took a few
seconds to see the new data (which is actually a reasonable way
to look at this).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>