Update nokogiri to version 1.16.5
This updates its packaged libxml2 to v2.12.7
to resolve what we believe are false claims that we are
vulnerable to CVE-2024-34459.

This is *NOT* believed to be a vulnerability in our
software. Nokogiri does not provide or expose the
code that is vulnerable (which is in xmllint).
Still, updating nokogiri will get rid of a false negative,
making it easier for us to detect real problems later.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>