Run "bundle update" on 2024-06-05
This addresses CVE-2024-28103 in actionpack.
This particular CVE is almost certainly not exploitable in our case,
as we don't directly depend on Permissions-Policy for security.
We use it, but only as a secondary additional defense.
However, we generally apply security updates unless there's a reason
to avoid it (as long as tests pass), in case our analysis is wrong.
This completes:
https://github.com/coreinfrastructure/best-practices-badge/security/dependabot/62

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>