Change user search to be OR instead of AND (#2190)
The new admin capability for user search is primarily for
GDPR requests (and similar). The previous implementation, if
provided with both name and email, returned only results that
met BOTH criteria (an "AND").

However, for GPDR requests, it's far more convenient to reply
all users matching EITHER criteria (an "OR"). We want to ensure
that we identify all possible records, even those that only
partly match.

This changes user search from "AND" to "OR", to better support
GDPR requests and similar.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>