Remove gem jquery-ui-rails
Remove gem jquery-ui-rails, eliminating vulnerability report
https://github.com/coreinfrastructure/best-practices-badge/dependabot/73

We previously updated gem jquery-ui-rails to version 7.0.0,
and had expected this to eliminate the vulnerability report.
Gem version 7.0.0 was released on 2024-03-13. This is a major
version number change, so we had expected it to take some extra
work but for it to also eliminate the vulnerability.

However, it turns out that the vulnerability report is for
versions <= 7.0.0, not < 7.0.0, as expected and as we
originally thought it said. Version 7.0.0 *is* the current version
of this gem as of 2024-11-27.

So we investigated our alternatives. It's likely to be unexploitable,
since it only affects a checkboxradio with HTML-like initial text
label. We limit where attackers can manipulate text, and escape it,
which makes this somewhat unlikely. Still, it was hard to be *certain*
there was no problem.

However, we *only* use this gem to implement a polyfill for Safari
to do an autocomplete of a datalist (FLOSS licenses).
The page https://caniuse.com/datalist indicates that support has
gotten *much* better, so we don't need the polyfill for typical
web browsers like Chrome, Firefox, Edge, Safari, and Opera.
Even if a user uses an unlikely browser for editing, this simply means
they would have to type in full text.

Therefore, we'll just remove the dependency entirely. Less code means
there's less code to maintain and less code to attack.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>