Switch to CircleCI Ruby 3.3.6 image for testing
Database migration now fails when running it from CircleCI.
Error message is from /usr/local/lib/heroku/bin/node
which asserts /lib/x86_64-linux-gnu/libc.so.6:
version `GLIBC_2.28' not found.

It appears that the heroku CLI requires a newer version of node,
which requires a newer glibc. Rather than try to force use of
an older version of node, let's try to update ruby and the rest.
We want to keep things relatively up to date.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>