Simplify using Google Chrome for testing via VCR
Make it simpler test the system with VCR and Google Chrome.

We don't want the system reaching out to the Internet for testing
in general. The system itself should work stand alone, and we don't
want its testing to interfere to with real systems.
We use VCR to record & play back interactions.

However, during testing we use Google Chrome. Chrome is *notoriously*
chatty and likes to update. We use Chrome-for-Testing to deal with
testing (a usual approach), but it likes to download information
from certain locations to handle updates.

These downloads are *not* part of the system's execution itself,
and we have no permissions to write to these facilities (so we can't
possibly interfere with their interactions). Thus, let's allow
these test-infrastructure downloads, so that we don't have to keep
trying to fix it up.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>