Remove extraneous parenthesis
While working on a company-internal fork of this project, I copied all
the criteria to a Markdown document to make it easier for us to assess
which criteria would need to change to suit our internal environment.
My editor highlighted this unmatched parenthesis.  I'm not sure if
removing it is the right thing to do, though, or if pulling the "see
also" sentence into the prior parenthetical statement would be better.

Signed-off-by: Jason M. Gates <jmgate@sandia.gov>