Update to Rails 7.1 (including related updates)
Update to Rails 7.1. This requires other related updates.

Oddly, 'rails-i18n' doesn't upgrade to the same number this time.
It *will* later when we upgrade again, so be careful about that.
I've added a spurious extra version number comment to hint at
that for later.

This updates paper_trail. However, it does *not* adjust the
older data in paper_trail. Paper_trail records data as it was,
and by default uses YAML. This is a problem due to a datatype change;
YAML records the *actual* datatype, even when we don't want to use it.
That should be addressed separately before merging this change.
This PR removes the known-failing test on papertrail.

This updates does *not* deal with a Rack deprecation.
We'll need to deal with that separately, but since it's still
supported, we can deal with it separately.
It will currently show:

> DEPRECATION WARNING: Setting action_dispatch.show_exceptions
> to false is deprecated. Set to :none instead.
> (called from call at ... /best-practices-badge/config/initializers/
> canonical_trailing_slash.rb:30)

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>