Change PaperTrail's storage from YAML to jsonb
We record old project data with PaperTrail, which by default stores
data in YAML.

YAML permits storing of arbitrary data types, which is
actually a bad thing because when Rails changes data types it
creates problems. This flexibility is possibly a security problem
in the long term if we deserialize things incorrectly.
Our serialization process is trusted, so that shouldn't be a problem,
but why not avoid it entirely long-term?

JSONB is more efficient to store and MUCH more efficient to query.
The PaperTail "versions" table is now our biggest table, so storing it
efficiently is a good idea.
It also doesn't allow arbitrary datatypes, which is a GOOD thing.

This should also make it easier to update to Rails 7.1, where
a datatype for a time with timezone causes problems.

Note: We don't configure PaperTrail, as it will automatically use
JSON if it's defined as jsonb in PostgreSQL. We might want to
configure it later just to ensure it stays in JSON even if we change
from PostgreSQL to a different database system.

To minimize production system downtime,
the database migration is *separate* from the process of converting
old data. The migration will immediately mean that new data will use
the new format. To convert old data, run:
`rake convert_papertrail_yaml_to_json`

After conversion is complete, we should eventually run a migration
that removes the old data column so it no longer takes up space:

~~~~
remove_column :versions, :old_yaml_object, :text
~~~~

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>