Remove deprecated `config/secrets.yml`
Rails' `config/secrets.yml` is deprecated and is not be
supported in Rails 7.2. Thus, we need to change how we
handle the `secret_key_base`.

After this commit, the application will continue to use
the environment variable `SECRET_KEY_BASE` when present
(this is how the `staging` and `production` sites already work).
When the environment variable is not set, the preset value in
the file `tmp/local_secret.txt` is used.

This *appears* to check in a private key and thus would be vulnerable,
but it's not a vulnerability in this case. The rationale for this
assertion is documented in the assurance case (as part of this
commit), as well as more details about why it's been done this way.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>