Update Rails 7.2->8.0; updates paper_trail
This updates to the current version of 8.0.*.
Rails 8.0.0 was released in 2024-11-07;
version 8.0.1 was released 2024-12-13.

Rails is our main framework, so it's helpful (though it takes
time) to keep up-to-date. A new X.0.0 version sometimes
has teething pains, but we now have a later version that
as far as I can tell resolves such pains.

This requires updating the gem paper_trail so we
update that as well.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>