Transition to Font Awesome 5
This completes the transition to Font Awesome 5.
In particular, it eliminates the language icon that unexpectedly
turned out to have some subtle licensing issues.
The names and how some icons are used changed, so we
had to change how we used them.

We are *only* using the free icons available under OSS licenses.

NOTE: This changes a few translations. Instead of embedding the
reference in the translation (which makes it difficult to change
how the icons are invoked), the translation *only* needs to
express the translation. This is easier for future translations, too.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>