Remove icons from translations
The translations were including special constructs to retrieve
icons from Font Awesome. In retrospect, that was a mistake.
They aren't translations; the icons stay the same in translations.
This also makes it harder to change how icons are retrieved
or which icons are used.

A previous commit inserted the icons and then the translations.
However, if we don't fix the translations, in many places a user
will see the icon twice (one from the app and one from the locale's
translation).

This fixes the translations so that they no longer include the icons.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>