Re-enable test of `BadPassword.force_load`
This re-enables the test of `BadPassword.force_load`.
If we loaded EVERYTHING, that would make that one test take a long time.
Instead, we'll add an optional parameter for how many entries to load,
and only load a few. That way we *can* test it without making
the test take forever.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>