Update favicon to current logo
This updates the favicon to the current logo stored in
"openssf_bestpracticesbadge.svg"

I used https://svg2ico.com/ to do this conversion (thank you!).

Icon (.ico) files can contain multiple sizes, and selecting sizes
nowadays is complicated. You can use 128 or 256 square, but these
result in big files, and we're expecting this to only appear
as part of a tab (so users won't see much of it).

I reviewed recommendations such as this one:
https://atlasiko.com/blog/web-development/favicon-size/

It's clear that 16, 32, and 48 pixel square are common cases.
Providing a 96 pixel square provides a decent desktop shortcut icon.
We could support larger pixels, up to 256 pixel square, but those
bloat the file for cases we don't expect to need. We want
first-use to be fast, so we'll do this.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>