Bugfix: Change external cache and cookie behavior
Add cases where cookies are set and where the external cache
is not saved.  We've been trying to optimize the website
by sometimes omitting cookies or aggressively caching.
However, this can sometimes cause some subtle bugs.

What *really* matters is that we use the CDN caches for badge images,
and that we not set cookies in that case (so that we avoid becoming
a tracking cookie). *That* is what's important.
Trying to be clever about cachine other pages that vary, or
not sending cookies in special cases, can lead to really complex
and hard-to-debug situations.

In general, we try to take the "simple" path, let's do that with
caches and cookies too.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>