Greatly simplify external cache handling
Eliminate external caching of most pages. Instead,
use an external cache (especially the CDN, but also the
user web broswer) only for the pages that *most*
matter for performance - particularly badge images.

Before we were using some clever tricks to externally
cache more pages. However, this turns out to be tricky to
get right because *many* pages vary depending on who is seeing them.
This includes the pages headers (which vary depending on whether or not
you're logged in), the "flash" reply (which may or may not
be in a cookie), and the CSRF countermeasures.

This really isn't *necessary*. Most legitimate retrievals are
badge images (because that's what is embedded in READMEs), so
optimizing that is enough.

For interactive use, we use fragment caching, which is built into
Rails and does a good job handling caches of *parts* of pages.

This simplifies the program, and thus makes it more likely
to be correct in the first palce.

Annoyingly, I can't find any way to convince Rails' built-in
CSRF mechanism to set the cache-control value to anything other than
`no-cache`, and that value still permits CDN caching.
So added a tiny `cache_control_fix.rb` shim to the rack middleware stack.
This shim changes any cache-control value that is *only* `no-cache` into
`private, no-cache` per our desires. Since this *always* runs after our
main code, it ensures that anything using the CSRF mechanism is *never*
cached by the CDN no matter what.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>