Remove pronto entirely (#2408)
* Remove pronto entirely

The "pronto" gem tries to run checks on "only what's relevant"
and can push comments onto a GitHub repo pull request.

However, it's failing. I think the GitHub API has changed.
What's more, I'm not convinced that this is helping.
We want to test everything before release, so a heuristic
to check only some things is dubious.
Also, there are concerns about giving a bot the right to
edit pull requests.

So let's simply always run the full test suite in CI/CD.
That way we're confident about the results, and don't need
to give the bot special permission. It's fewer dependencies
to keep up to date (and interfere with other updates), too.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>