Add a project .mdlrc file
By adding a .mdlrc file, we make it easier to directly invoke
markdownlint (mdl) for specific files, because the .mdlrc file
tells the system where to load our style.

This is handy when we have a problem with a specific file or set of files.
we can run mdl directly on just those files with just `mdl FILENAMES`,
without having to carefully re-specify our style on the command line.

This simplified our rake file, and we may as well document why we
use "bundle exec" there even though it's not necessary in the
development environment if configured as recommended.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>