Ignore ,temporary files in linters; AI assistants
AI assistants like to create temporary files to investigate
or validate constructs. This is good for helping to
counter hallucinations or errors.

However, this creates a lot of junk that can be a pain
to remove. Even worse, it causes linters to fail; the
temporary code often doesn't match our stringent style guides.
There's no need to enforce our style code on temporary code
we're about to throw away.

So we'll formally use an old Unix convention that I believe many
have used: files beginning with comma are temporary files.
The advantage of this convention is that's really easy to
identify the files, and their suffix will still correctly
tell you the filetype.

We'll now modify the rake tasks so the various linters will
ignore these filenames. Now the AI code assistants and the humans
can create temporary files to investigate issues without having
to constantly make the linters happy.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>