Add rake task percent_gems_up_to_date
Add rake task percent_gems_up_to_date, and call it in the
default rake process.

This new task only does anything useful if there's a network
connection. If there is, it reports the number of gems
out-of-date, gems current, total gems, and % gems up-to-date.
This gives us quick info on how up-to-date we are without
a lot of text output.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>