Simplify scope declarations (#2434)
* Simplify scope declarations

More recent versions of Rails, combined with more recent versions
of Ruby, enable a simpler syntax for defining scopes, replacing
5 lines of code with 1 in each case. We have several scopes to define,
so let's switch to the simpler syntax.

We also add a number of unit tests to make *sure* nothing changed.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update CLAUDE.md

Rubocop's -A option (unsafe autocorrect) is much trouble than it's
worth. It's better to let the AI fix it than to run the unsafe
autocorrect.

* Remove bad cop in .rubocop.yml, fix projects controller

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>