Limit number of repos loaded for a user
* Limit number of repos from github_user_projects

Some users have a vast number of repos that they own or
have control over in some way, which can lead to rediculous results.
We had one user where this caused a response to take 28 seconds
(instead of milliseconds) and all by *itself* this request
increased memory use by 457MB.

This caps the number of repos we consider to a smaller
reasonable number. We sort by 'updated', so
the repos that are being actively maintained will be the ones
returned if the cap is being used.

In extreme cases like this, the expected impact of this change is:

- Memory per user page view: 457MB -> ~0.2MB (99.96% reduction)
- Response time: 25-28 seconds -> less than 1 second
- GitHub API calls: 30-100+ -> 1

* Only consider public GitHub projects

Change so we only consider public GitHub projects
('public' not 'all) when creating a list of user projects.

* Only list public repos when creating new entry

* Fix reports from GitHub Copilot and spellcheck

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>