Rename is_disabled to view_only
Many UI components look different when they allow editing vs.
when they *disable* editing. So we've been using "is_disabled"
to indicate the mode that only permits viewing (not editing).
However, in higher-level constructs this is confusing
(what is being disabled?).

This changes the name of the boolean to "view_only".
It still has the same sense ("true" means "only viewing no editing),
but giving its name a higher-level meaning makes it easier to follow.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>