Highlight current section in list of sections
The server-side code already indicated the "active" section,
but previously there was no style that made it visible.

This adds styling so the current (active) section is highlighted,
in this case by a colored checkmark. Exactly the same text is
transparently before the other sections to ensure that the text
alignment is perfect.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>