Rename SCSS partial filenames per convention
SCSS partials are supposed to have filenames beginning
with "_", but we didn't do that consistently.
Rename to make them consistent.

The "application.scss" becomes "application.css" which
imports everything else. This single CSS is cached on the
user's browser, so once a user loads it, they won't need
to load it again. We don't use the other files as anything
other than partials, so rendering them so they can be
used independently is a waste of time.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>