Baseline phase2 (#2527)
* Update phase 2 plan - hybrid approach

We're trying to minimize changes to the code and want to ensure
that we can translate the results to other natural languages
(while not requiring translators to do unnecessary work).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Phase 2 heart: Pull baseline and migrate.

This also improves fix_markdown.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Minor fixups to script/fix_markdown

The script/fix_markdown script isn't ready for mass use.
Make that clear, and do some fixups of it so that maybe
someday it will be.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix markdownlint warnings

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix link from baseline_plan.md

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* baseline criteria: met_url_required: false

For now, let's *not* require URLs for any baseline answers.
That will make it easy to use as we get started.
We can change this later.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix tests for phase 2

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Make masssive changes to machine translation plan

The old plan for machine translation was terrible.
Here's a better one.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix formatting issues

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix markdown processor thread-safety

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Don't call markdown processor for trivial cases

As an optimization, detect trivial strings & don't
call the full markdown processor to process them.
We expect to have many cases where only trivial strings are
provided.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Don't consider test coverage of non-prod code

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Fix generating quotes

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update codecov configuration

Fix issues and clean up the configuration for
code coverage reporting.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Tweak machine translation plan

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Translation: Use YAML, improve context

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Improve using existing translations as start

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

---------

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>