Check on first URL that assets were precompiled
Several times I *and* the AI have modified a precompiled
asset and forgotten to precompile it by running
`rake assets:precompile`. This often leads to
long, discouraging, and wasteful debugging sessions.

This teaches to system to check if we forgot to precompile assets.
At first I had a complicated check,
but there's no need for that; if there's an source asset that's
newer than a compiled asset, we forgot to do it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>