Improve testability of projects controller
SimpleCov can't detect when branches are run on class
initialization. So let's reorganize things to pull a branch
into its own class method, and create tests for it.

This means we make it easier to test, improve its testing, and
make it clear to our test infrastructure that we *are* testing
both the true and false branches.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>