redirect_to_default_section: don't load project
Redirection of the project doesn't need to load all
the project data. Let's avoid loading what we can.
Currently we can *always* avoid loading it. Eventually
we'll consult the project for its default section, but
even then, we won't load most of the project data just
to redirect to the section.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>