Add defensive measures if a status is a string
Add defensive measures so that displays work even if a
criterion's status is a string internally.

We've converted all "status" values from strings to small integers
internally everywhere. This only impacts how the data is stored
and processed internally; external viewers see the more meaningful
strings. This greatly optimizes processing.

However, an initial push to staging caused problems, because it was
trying convert a string instead of the expected int. I believe it was
from a stale dyno cache (restarting appears to have fixed the problem),
but it's impossible to reproduce the problem. We want the system to
work reliably.

This introduces two small defensive measures, so that if the software
somehow ends up with a string internally, it'll quietly pass it out
to the outside world. This means it'll appear to work just fine
to the outside world.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>